making custom evals configurable
